dom/base/test/chrome/test_bug421622.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=421622
-->
<window title="Mozilla Bug 421622"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=421622"
     target="_blank">Mozilla Bug 421622</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

    /** Test for Bug 421622 **/
    const SJS_URL = "http://mochi.test:8888/tests/dom/base/test/chrome/bug421622-referer.sjs";
    const REFERER_URL = "http://www.mozilla.org/";

    var req = new XMLHttpRequest();
    req.open("GET", SJS_URL, false);
    req.setRequestHeader("Referer", REFERER_URL);
    req.send(null);

    is(req.responseText,
       "Referer: " + REFERER_URL,
       "Referer header received by server does not match what was set");
  ]]></script>
</window>
